carrinho/requirements.txt
blinker==1.4
dj-database-url==0.4.1
Django==1.10.2
django-cors-headers==1.3.0
django-filter==0.15.2
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.4.7
gunicorn==19.6.0
Markdown==2.6.7
mongoengine==0.9.0
Pygments==2.1.3
pymongo==2.8
requests==2.11.1
whitenoise==3.2.2
